Remove fake strategy from mining director tests (#278)
All these tests didn't use our real code, they used a fake strategy
instead of testing our real code code. We're trying to test our real
code, and also remove FakeStrategy and generics all together.
